Moved vending machine functions into different vending_machine_utils.py
